Implemented UserModelWithName model.
