+++ src/main/java/com/smartsheet/api/models/UserProfile.java
@@ -23,7 +23,7 @@ package com.smartsheet.api.models;
 /**
  * A profile object that contains the basic fields that most profiles will contain.
  */
-public class UserProfile extends AbstractUser {
+public class UserProfile extends UserModel {
 	/**
 	 * Represents the user's time zone
 	 */
